--- app/src/main/java/com/finoldigital/ygolp/presentation/CalculatorMode.kt
@@ -1,7 +1,0 @@
-package com.finoldigital.ygolp.presentation
-
-object CalculatorMode {
-    const val SET = 0
-    const val SUBTRACT = 1
-    const val ADD = 2
-}
