+++ app/src/main/java/com/finoldigital/ygolp/presentation/screens/CalculatorMode.kt
@@ -1,0 +1,11 @@
+package com.finoldigital.ygolp.presentation.screens
+
+enum class CalculatorMode(val value: Int) {
+    SET(0),
+    SUBTRACT(1),
+    ADD(2);
+
+    companion object {
+        fun fromInt(value: Int) = entries.find { it.value == value } ?: SET
+    }
+}
