+++ app/src/main/java/com/finoldigital/ygolp/presentation/screens/Screen.kt
@@ -1,4 +1,4 @@
-package com.finoldigital.ygolp.presentation
+package com.finoldigital.ygolp.presentation.screens
 
 sealed class Screen(val route: String) {
     object LifePoints : Screen("lifepoints/{player}") {
@@ -6,6 +6,6 @@ sealed class Screen(val route: String) {
     }
 
     object Calculator : Screen("calculator/{player}/{initialCalculatorMode}") {
-        fun createRoute(player: Int, mode: Int) = "calculator/$player/$mode"
+        fun createRoute(player: Int, mode: CalculatorMode) = "calculator/$player/${mode.value}"
     }
 }
